+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## Added
+- Invalid variance error
+
+## Changed
+- Changed start fee rate for both spot/perp in voting constants (topic `Fee Rate`)
+
 ## [v2.3] - 2025-11-05
 
 ## Added
 - Rust docs for models
 - Rust docs for instruction data
 
-## Changed
-- Changed start fee rate for both spot/perp in voting constants (topic `Fee Rate`)